add custom hook useProposalURLs
