+++ src/containers/Proposal/hooks/index.js
@@ -1,2 +1,3 @@
 export { default as useProposalVote } from "./useProposalVote";
 export { default as useProposalsBatch } from "./useProposalsBatch";
+export { default as useProposalURLs } from "./useProposalURLs";
